+++ src/module.test.ts
@@ -143,7 +143,7 @@ describe('TestPlatform', () => {
 
     await dynamicPlatform.onStart('Test reason');
     expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'onStart called with reason:', 'Test reason');
-    expect(addBridgedEndpointSpy).toHaveBeenCalledTimes(62);
+    expect(addBridgedEndpointSpy).toHaveBeenCalledTimes(63);
   });
 
   it('should execute the commandHandlers', async () => {