+++ CHANGELOG.md
@@ -23,6 +23,20 @@ If you like this project and find it useful, please consider giving it a star on
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="120"></a>
 
+## [2.0.7] - 2026-01-31
+
+### Added
+
+- [SmartOutlet]: Added a smart outlet with an energy meter and four individually controllable sockets (with tagList 1, 2, 3 and 4).
+
+### Changed
+
+- [package]: Updated dependencies.
+- [package]: Required matterbridge 3.5.2.
+- [package]: Bumped package to automator v.3.0.3.
+
+<a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>
+
 ## [2.0.6] - 2026-01-31
 
 ### Added